Correct sizeof vs int compare.
git-svn-id: file:///home/jj/hercules.svn/trunk@6189 956126f8-22a0-4046-8f4a-272fa8102e63

+++ hsccmd.c
@@ -4154,7 +4154,7 @@ int devlist_cmd(int argc, char *argv[], char *cmdline)
         {
             int i;
             strcpy( devtype, argv[1] );
-            for ( i = 0; i < strlen( devtype ); i++ )
+            for ( i = 0; i < (int)strlen( devtype ); i++ )
                 if ( isalpha( devtype[i] ) )
                     devtype[i] = toupper( devtype[i] );
         }
